refactor(memory): remove deprecated observation tools and split neo4j adapter
