+++ crates/mm-core/src/lib.rs
@@ -14,19 +14,16 @@ mod root;
 pub use error::{CoreError, CoreResult};
 pub use mm_memory::MemoryService;
 pub use operations::{
-    AddObservationsCommand, AddObservationsResult, CreateEntitiesCommand, CreateEntitiesResult,
-    CreateRelationshipsCommand, CreateRelationshipsResult, DeleteEntitiesCommand,
-    DeleteEntitiesResult, DeleteRelationshipsCommand, DeleteRelationshipsResult,
-    FindEntitiesByLabelsCommand, FindEntitiesByLabelsResult, FindEntitiesByLabelsResultType,
-    FindRelationshipsCommand, FindRelationshipsResult, FindRelationshipsResultType,
-    GetEntityCommand, GetEntityResult, GetProjectContextCommand, GetProjectContextResult,
-    ListProjectsCommand, ListProjectsResult, ProjectFilter, RemoveAllObservationsCommand,
-    RemoveAllObservationsResult, RemoveObservationsCommand, RemoveObservationsResult,
-    SetObservationsCommand, SetObservationsResult, UpdateEntityCommand, UpdateEntityResult,
-    UpdateRelationshipCommand, UpdateRelationshipResult, add_observations, create_entities,
-    create_relationships, delete_entities, delete_relationships, find_entities_by_labels,
-    find_relationships, get_entity, get_project_context, list_projects, remove_all_observations,
-    remove_observations, set_observations, update_entity, update_relationship,
+    CreateEntitiesCommand, CreateEntitiesResult, CreateRelationshipsCommand,
+    CreateRelationshipsResult, DeleteEntitiesCommand, DeleteEntitiesResult,
+    DeleteRelationshipsCommand, DeleteRelationshipsResult, FindEntitiesByLabelsCommand,
+    FindEntitiesByLabelsResult, FindEntitiesByLabelsResultType, FindRelationshipsCommand,
+    FindRelationshipsResult, FindRelationshipsResultType, GetEntityCommand, GetEntityResult,
+    GetProjectContextCommand, GetProjectContextResult, ListProjectsCommand, ListProjectsResult,
+    ProjectFilter, UpdateEntityCommand, UpdateEntityResult, UpdateRelationshipCommand,
+    UpdateRelationshipResult, create_entities, create_relationships, delete_entities,
+    delete_relationships, find_entities_by_labels, find_relationships, get_entity,
+    get_project_context, list_projects, update_entity, update_relationship,
 };
 pub use ports::Ports;
 pub use root::{Root, RootCollection};
