+++ crates/mm-core/src/operations/mod.rs
@@ -1,6 +1,5 @@
 mod common;
 
-pub mod add_observations;
 pub mod create_entity; // renamed file but keep mod
 pub mod create_relationship; // rename later
 pub mod delete_entities;
@@ -10,13 +9,9 @@ pub mod find_relationships;
 pub mod get_entity;
 pub mod get_project_context;
 pub mod list_projects;
-pub mod remove_all_observations;
-pub mod remove_observations;
-pub mod set_observations;
 pub mod update_entity;
 pub mod update_relationship;
 
-pub use add_observations::{AddObservationsCommand, AddObservationsResult, add_observations};
 pub use create_entity::{CreateEntitiesCommand, CreateEntitiesResult, create_entities};
 pub use create_relationship::{
     CreateRelationshipsCommand, CreateRelationshipsResult, create_relationships,
@@ -38,13 +33,6 @@ pub use get_project_context::{
     GetProjectContextCommand, GetProjectContextResult, ProjectFilter, get_project_context,
 };
 pub use list_projects::{ListProjectsCommand, ListProjectsResult, list_projects};
-pub use remove_all_observations::{
-    RemoveAllObservationsCommand, RemoveAllObservationsResult, remove_all_observations,
-};
-pub use remove_observations::{
-    RemoveObservationsCommand, RemoveObservationsResult, remove_observations,
-};
-pub use set_observations::{SetObservationsCommand, SetObservationsResult, set_observations};
 pub use update_entity::{UpdateEntityCommand, UpdateEntityResult, update_entity};
 pub use update_relationship::{
     UpdateRelationshipCommand, UpdateRelationshipResult, update_relationship,
