+++ crates/mm-memory-neo4j/src/adapters/neo4j/mod.rs
@@ -1,0 +1,6 @@
+mod config;
+mod helpers;
+mod repository;
+
+pub use config::Neo4jConfig;
+pub use repository::Neo4jRepository;
