+++ crates/mm-server/src/mcp/mod.rs
@@ -1,6 +1,5 @@
 #[macro_use]
 mod macros;
-pub mod add_observations; // deprecated
 pub mod create_entities;
 pub mod create_relationships;
 pub mod delete_entities;
@@ -11,15 +10,11 @@ pub mod find_relationships;
 pub mod get_entity;
 pub mod get_project_context;
 pub mod list_projects;
-pub mod remove_all_observations;
-pub mod remove_observations;
-pub mod set_observations;
 pub mod update_entity;
 pub mod update_relationship;
 
 use rust_mcp_sdk::tool_box;
 
-pub use add_observations::AddObservationsTool;
 pub use create_entities::CreateEntitiesTool;
 pub use create_relationships::CreateRelationshipsTool;
 pub use delete_entities::DeleteEntitiesTool;
@@ -29,9 +24,6 @@ pub use find_relationships::FindRelationshipsTool;
 pub use get_entity::GetEntityTool;
 pub use get_project_context::GetProjectContextTool;
 pub use list_projects::ListProjectsTool;
-pub use remove_all_observations::RemoveAllObservationsTool; // deprecated
-pub use remove_observations::RemoveObservationsTool; // deprecated
-pub use set_observations::SetObservationsTool; // deprecated
 pub use update_entity::UpdateEntityTool;
 pub use update_relationship::UpdateRelationshipTool;
 
@@ -46,11 +38,6 @@ tool_box!(
         FindEntitiesByLabelsTool,
         FindRelationshipsTool,
         GetEntityTool,
-        // Deprecated observation tools below
-        SetObservationsTool,
-        AddObservationsTool,
-        RemoveAllObservationsTool,
-        RemoveObservationsTool,
         GetProjectContextTool,
         ListProjectsTool,
         UpdateEntityTool,
