--- crates/mm-server/src/mcp/remove_all_observations.rs
@@ -1,67 +1,0 @@
-use mm_core::{RemoveAllObservationsCommand, remove_all_observations};
-use mm_utils::IntoJsonSchema;
-use rust_mcp_sdk::macros::mcp_tool;
-use serde::{Deserialize, Serialize};
-
-#[mcp_tool(
-    name = "remove_all_observations",
-    description = "Remove all observations from an entity"
-)]
-#[derive(Debug, Serialize, Deserialize, schemars::JsonSchema)]
-pub struct RemoveAllObservationsTool {
-    pub name: String,
-}
-
-impl RemoveAllObservationsTool {
-    generate_call_tool!(
-        self,
-        RemoveAllObservationsCommand { name },
-        "All observations removed from '{}'",
-        remove_all_observations
-    );
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use mm_core::Ports;
-    use mm_memory::{MemoryConfig, MemoryError, MemoryService, MockMemoryRepository};
-    use std::sync::Arc;
-
-    #[tokio::test]
-    async fn test_call_tool_success() {
-        let mut mock = MockMemoryRepository::new();
-        mock.expect_remove_all_observations()
-            .withf(|name| name == "test:entity")
-            .returning(|_| Ok(()));
-
-        let service = MemoryService::new(mock, MemoryConfig::default());
-        let ports = Ports::new(Arc::new(service));
-
-        let tool = RemoveAllObservationsTool {
-            name: "test:entity".to_string(),
-        };
-
-        let result = tool.call_tool(&ports).await.expect("tool should succeed");
-        let text = result.content[0].as_text_content().unwrap().text.clone();
-        assert_eq!(text, "All observations removed from 'test:entity'");
-    }
-
-    #[tokio::test]
-    async fn test_call_tool_repository_error() {
-        let mut mock = MockMemoryRepository::new();
-        mock.expect_remove_all_observations()
-            .withf(|name| name == "test:entity")
-            .returning(|_| Err(MemoryError::runtime_error("fail")));
-
-        let service = MemoryService::new(mock, MemoryConfig::default());
-        let ports = Ports::new(Arc::new(service));
-
-        let tool = RemoveAllObservationsTool {
-            name: "test:entity".to_string(),
-        };
-
-        let result = tool.call_tool(&ports).await;
-        assert!(result.is_err());
-    }
-}
